fastapi_jwt_rbac/requirements.txt
fastapi
sqlmodel
uvicorn
python-jose[cryptography]
passlib[bcrypt]
python-multipart
psycopg2-binary
